requirements.txt
paho-mqtt
pyserial-asyncio
